Move off of bitnami
Move off of bitnami image for gitlab CI, use `alpine/k8s` instead. This image is recommended by the Diamond Cloud Team. We have tested it internally on a beamline (b01-1).

Note that alpine/k8s has no `latest` tag so we will need to keep it up to date in this template.